@@ -1,5 +1,5 @@
 let postcssSvg = require('postcss-inline-svg')({
-    paths: [__dirname + '/resources/img']
+    paths: [__dirname + '/src/img']
 });
 
 module.exports = {
