@@ -1,22 +1,20 @@
-import Vue from 'vue';
-
 import KyoMedias from "./medias/KyoMedias";
-Vue.component(KyoMedias.name, KyoMedias);
+kyoto.component(KyoMedias.name, KyoMedias);
 
 import KyoMediaCreate from "./medias/KyoMediaCreate";
-Vue.component(KyoMediaCreate.name, KyoMediaCreate);
+kyoto.component(KyoMediaCreate.name, KyoMediaCreate);
 
 import KyoMediaEdit from "./medias/KyoMediaEdit";
-Vue.component(KyoMediaEdit.name, KyoMediaEdit);
+kyoto.component(KyoMediaEdit.name, KyoMediaEdit);
 
 import KyoVideos from "./modals/KyoVideos";
-Vue.component(KyoVideos.name, KyoVideos);
+kyoto.component(KyoVideos.name, KyoVideos);
 
 import KyoUploads from "./modals/KyoUploads";
-Vue.component(KyoUploads.name, KyoUploads);
+kyoto.component(KyoUploads.name, KyoUploads);
 
 import KyoMediaSelect from "./inputs/KyoMediaSelect";
-Vue.component(KyoMediaSelect.name, KyoMediaSelect);
+kyoto.component(KyoMediaSelect.name, KyoMediaSelect);
 
 if ( console && console.log ) {
     console.log('kyoto/media ready.');
